version: Use uint32 for major/minor/patch.
This modifies the semver handling to specifically use uint32 instead of
architecture-dependent uint so bounds checking is consistent across
architectures and only needs to be done in one place.

+++ internal/rpcserver/rpcserverhandlers_test.go
@@ -8551,9 +8551,9 @@ func TestHandleVersion(t *testing.T) {
 		},
 		"dcrd": {
 			VersionString: version.String(),
-			Major:         uint32(version.Major),
-			Minor:         uint32(version.Minor),
-			Patch:         uint32(version.Patch),
+			Major:         version.Major,
+			Minor:         version.Minor,
+			Patch:         version.Patch,
 			Prerelease:    version.PreRelease,
 			BuildMetadata: buildMeta,
 		}}
